GuiUncraftingTable moved to client package
